Make Math with CGFloat generic
CGFloat is 32 or 64bit depending on build system.
need to use generic math (which requires, due to a compiler bug,
disabling modules)
http://stackoverflow.com/questions/23333287/tgmath-h-doesnt-work-if-modu
les-are-enabled

+++ Observer/AKRDirectionIndicatorView.m
@@ -23,7 +23,7 @@
 {
     if (fabs(azimuth - _azimuth) < 0.02)
         return;
-    float oldAzimuth = _azimuth;
+    CGFloat oldAzimuth = _azimuth;
     _azimuth = azimuth;
     if (2.0 < (fabs(azimuth - oldAzimuth))) {
         [self setNeedsDisplay];
